+++ src/commands/chat/format.ts
@@ -1,7 +1,7 @@
 import { AssistantResponse } from '@callstack/byorg-core';
+import { colorAssistant, colorVerbose } from '../../colors.js';
 import { formatSpeed, formatTime } from '../../format.js';
-import { colorAssistant, colorVerbose } from './colors.js';
-import { getVerbose } from './output.js';
+import { getVerbose } from '../../output.js';
 
 export function formatResponse(response: AssistantResponse) {
   let result = colorAssistant(response.content);
